requirements.txt
#requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
python-nmap>=0.7.1
colorama>=0.4.6
urllib3>=2.0.0
cryptography>=40.0.0
python-dateutil>=2.8.2
